requirements.txt
flask==2.0.1
twitchio==2.6.0
python-dotenv==0.19.0
websockets==10.1